Move cmd/worker/memo into internal.
